clnt_vc_control: fix VERS/PROG size
Commit fac3eb51 converted some operations to use memcpy, but the size
of copied areas is now u_long, previously it was u_int32_t.

This broke mount on ppc64/s390x:
  # ./mount.nfs -vvv -onfsvers=3,tcp rhel6-nfs:/export/home /mnt/test
  mount.nfs: timeout set for Fri May  3 15:20:14 2013
  mount.nfs: trying text-based options 'nfsvers=3,tcp,addr=10.1.1.5'
  mount.nfs: prog 100003, trying vers=3, prot=6
  mount.nfs: portmap query failed: RPC: Program/version mismatch
  mount.nfs: Protocol not supported

+++ src/clnt_vc.c
@@ -505,7 +505,7 @@ clnt_vc_control(cl, request, info)
 	sigset_t newmask;
 	int rpc_lock_value;
 	u_int32_t tmp;
-	u_long ltmp;
+	u_int32_t ltmp;
 
 	assert(cl != NULL);
